+++ BlueprintUI/Sources/Layout/ConstrainedAspectRatio.swift
@@ -4,31 +4,14 @@ import UIKit
 public struct ConstrainedAspectRatio: Element {
     /// Represents how the content should size itself relative to its parent.
     public enum ContentMode: Equatable {
-        /// The content will be sized to fill its parent while
-        /// maintaining the specified aspect ratio.
-        ///
-        /// In ``LayoutMode/legacy`` layout mode, if the parent is unconstrained in all dimensions,
-        /// the content size will be used for measurement and will behave like `fitContent`. If the
-        /// parent is unconstrained in one dimension, the element will fill the constrained
-        /// dimension.
-        ///
-        /// In ``LayoutMode/caffeinated``, if the parent is unconstrained in any dimension, this
-        /// element will report a size of `infinity`.
-        ///
-        @available(*, deprecated, message: "Limited utility in Caffeinated Layout")
-        case fillParent
-
         /// The content will be sized to fit within its parent while maintaining the specified
         /// aspect ratio.
         ///
         /// If the parent is unconstrained in one dimension, the element will fit the constrained
         /// dimension.
         ///
-        /// In ``LayoutMode/legacy`` layout mode, if the parent is unconstrained in all dimensions,
-        /// the content size will be used for measurement and will behave like `fitContent`.
-        ///
-        /// In ``LayoutMode/caffeinated`` layout mode, if the parent is unconstrained in both
-        /// dimensions, this element will report a size of `infinity`.
+        /// If the parent is unconstrained in both dimensions, this element will report a size of
+        /// `infinity`.
         ///
         case fitParent
 
@@ -55,8 +38,7 @@ public struct ConstrainedAspectRatio: Element {
             let availableWidth: CGFloat
 
             switch self {
-            case .fillParent,
-                 .fitParent:
+            case .fitParent:
                 availableHeight = constraint.height.constrainedValue ?? contentSize.height
                 availableWidth = constraint.width.constrainedValue ?? contentSize.width
             case .fitContent,
@@ -69,26 +51,6 @@ public struct ConstrainedAspectRatio: Element {
             let constrainedWidth = aspectRatio.width(forHeight: availableHeight)
 
             switch self {
-            case .fillParent:
-                if constraint.width.constrainedValue == nil &&
-                    constraint.height.constrainedValue == nil
-                {
-                    if constrainedWidth > availableWidth {
-                        return aspectRatio.size(forWidth: constrainedWidth)
-                    } else if constrainedHeight > availableHeight {
-                        return aspectRatio.size(forHeight: constrainedHeight)
-                    }
-                    return contentSize
-                } else if constraint.width.constrainedValue == nil {
-                    return aspectRatio.size(forWidth: constrainedWidth)
-                } else if constraint.height.constrainedValue == nil {
-                    return aspectRatio.size(forHeight: constrainedHeight)
-                } else if constrainedWidth < availableWidth {
-                    return aspectRatio.size(forHeight: constrainedHeight)
-                } else {
-                    return aspectRatio.size(forWidth: constrainedWidth)
-                }
-
             case .fitParent:
                 if constraint.width.constrainedValue == nil &&
                     constraint.height.constrainedValue == nil
@@ -134,24 +96,6 @@ public struct ConstrainedAspectRatio: Element {
         ) -> CGSize {
 
             switch self {
-            case .fillParent:
-                switch (proposal.width, proposal.height) {
-
-                case (.atMost(let maxWidth), .atMost(let maxHeight)):
-                    let constrainedWidth = aspectRatio.width(forHeight: maxHeight)
-
-                    if constrainedWidth > maxWidth {
-                        // grow width to match height
-                        return aspectRatio.size(forHeight: maxHeight)
-                    } else {
-                        // grow height to match width
-                        return aspectRatio.size(forWidth: maxWidth)
-                    }
-
-                default:
-                    return .infinity
-                }
-
             case .fitParent:
                 switch (proposal.width, proposal.height) {
                 case (.unconstrained, .unconstrained):
@@ -234,19 +178,6 @@ public struct ConstrainedAspectRatio: Element {
     private struct Layout: SingleChildLayout {
         var aspectRatio: AspectRatio
         var contentMode: ContentMode
-
-        func measure(in sizeConstraint: SizeConstraint, child: Measurable) -> CGSize {
-            let contentSize = child.measure(in: sizeConstraint)
-            return contentMode.constrain(
-                contentSize: contentSize,
-                in: sizeConstraint,
-                to: aspectRatio
-            )
-        }
-
-        func layout(size: CGSize, child: Measurable) -> LayoutAttributes {
-            LayoutAttributes(size: size)
-        }
 
         func sizeThatFits(
             proposal: SizeConstraint,